requirements.txt
absl-py==1.3.0
adam==0.0.0.dev0
altgraph==0.17.4
appdirs==1.4.4
asgiref==3.5.1
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.9.24
cffi==1.16.0
charset-normalizer==2.1.1

colorama==0.4.6
configparser==4.0.2
contourpy==1.2.1
crispy-bootstrap4==2022.1
cryptography==42.0.5
cycler==0.12.1
distlib==0.3.6
dj-database-url==2.2.0
Django==4.0.4
django-background-tasks==1.2.8
django-crispy-forms==2.0
django-otp==1.5.1
django-phonenumber-field==7.1.0
django-recaptcha==3.0.0
filelock==3.8.0
flatbuffers==22.11.23
fonttools==4.51.0
fsspec==2024.6.1
gast==0.4.0
google-api-core==2.18.0
google-api-python-client==2.124.0
google-auth==2.15.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.63.0
graphviz==0.20.3
grpcio==1.51.1
gunicorn==23.0.0

hexdump==3.3
httplib2==0.22.0
huggingface-hub==0.23.5
idna==3.4
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.0
keras==2.11.0
kiwisolver==1.4.5

libclang==14.0.6
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.4
mkl==2021.4.0
more-itertools==5.0.0
mpmath==1.3.0
networkx==3.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==24.0
pandas==1.5.2
pefile==2023.2.7
phonenumbers==8.13.17
pillow==10.3.0
platformdirs==2.5.2
proto-plus==1.23.0
protobuf==3.19.6
psutil==5.9.8
psycopg2==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.13.1
pycparser==2.21
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pyparsing==2.4.7
PyPDF2==3.0.1

pyshark==0.6
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-evtx==0.7.4
python-http-client==3.3.7
pytz==2022.6
PyYAML==6.0.1
regex==2024.5.15
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.3
scapy==2.5.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sendgrid==3.6.5
sendgrid-django==4.2.0
shiboken6==6.6.1
six==1.16.0
sqlparse==0.4.2
sympy==1.12
tbb==2021.13.0

tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1

termcolor==2.1.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.3.1
tqdm==4.66.4
transformers==4.42.4
typing_extensions==4.12.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.13
virtualenv==20.16.5
watchdog==4.0.0
Werkzeug==2.2.2
whitenoise==6.6.0

wrapt==1.14.1
zipp==1.0.0
